# internal/domain/position.go
package domain

import (
	"time"

	"github.com/google/uuid"
)

type Position struct {
	ID        uuid.UUID `json:"id"`
	Name      string    `json:"name"`
	IsActive  bool      `json:"is_active"`
	CreatedAt time.Time `json:"created_at"`
}